"fix bronze layer to be suited to streaming case"
